Extracts SensorApi interface.
* Prepares for addition of InfluxDb
