+++ cellar-app/src/main/java/nl/lengrand/cellar/store/faunadb/FaunaDbSetup.java
@@ -1,11 +1,11 @@
-package nl.lengrand.cellar.faunadb;
+package nl.lengrand.cellar.store.faunadb;
 import com.faunadb.client.FaunaClient;
 import com.faunadb.client.types.Value;
 
 import java.util.concurrent.ExecutionException;
 
 import static com.faunadb.client.query.Language.*;
-import static nl.lengrand.cellar.faunadb.Connection.*;
+import static nl.lengrand.cellar.store.faunadb.Connection.*;
 
 
 /*
