+++ cellar-app/src/main/java/nl/lengrand/cellar/store/faunadb/SensorApi.java
@@ -1,0 +1,7 @@
+package nl.lengrand.cellar.store.faunadb;
+
+import nl.lengrand.cellar.SensorValue;
+
+public interface SensorApi {
+    void add(SensorValue values);
+}
